fix(suite): add missing 'use device here' button in the device settings
